requirements.txt
numpy
pyserial
tqdm
logging